refactor(query): remove uesless code
