+++ src/query/storages/fuse/src/pruning/mod.rs
@@ -15,10 +15,9 @@
 mod block_pruner;
 mod bloom_pruner;
 mod fuse_pruner;
-mod new_segment_pruner;
+mod segment_pruner;
 mod pruner_location;
 mod pruning_statistics;
-mod segment_pruner;
 
 pub use block_pruner::BlockPruner;
 pub use bloom_pruner::BloomPruner;
@@ -28,4 +27,3 @@ pub use fuse_pruner::PruningContext;
 pub use pruner_location::create_segment_location_vector;
 pub use pruner_location::SegmentLocation;
 pub use pruning_statistics::FusePruningStatistics;
-pub use segment_pruner::SegmentPruner;
